lpi2c: imx8: fix i2c configuration when using high clock-rate
Some internal variables were overflowing when using a high clock-rate
like 48MHz and low clock-frequency like 100kbps.

https://jira.digi.com/browse/DEL-6269

(cherry picked from commit 1af0c9ba04f3bd97ad45be660c6ea19284864eac)

+++ drivers/i2c/busses/i2c-imx-lpi2c.c
@@ -222,7 +222,7 @@ static void lpi2c_imx_stop(struct lpi2c_imx_struct *lpi2c_imx)
 /* CLKLO = I2C_CLK_RATIO * CLKHI, SETHOLD = CLKHI, DATAVD = CLKHI/2 */
 static int lpi2c_imx_config(struct lpi2c_imx_struct *lpi2c_imx)
 {
-	u8 prescale, filt, sethold = 0, clkhi, clklo, datavd;
+	u16 prescale, filt, sethold = 0, clkhi, clklo, datavd;
 	unsigned int clk_rate, clk_cycle;
 	enum lpi2c_imx_pincfg pincfg;
 	unsigned int temp;
